refactor(#334): 신고하기 컴포넌트 공통화
